add support for API "envelopes"
