+++ src/lib.rs
@@ -1,4 +1,6 @@
 mod api_client;
 mod serde;
+mod types;
 
 pub use api_client::*;
+pub use types::*;
